+++ tests/unittests/classification/test_confusion_matrix.py
@@ -151,7 +151,9 @@ class TestMulticlassConfusionMatrix(MetricTester):
             preds=preds,
             target=target,
             metric_class=MulticlassConfusionMatrix,
-            reference_metric=partial(_sklearn_confusion_matrix_multiclass, normalize=normalize, ignore_index=ignore_index),
+            reference_metric=partial(
+                _sklearn_confusion_matrix_multiclass, normalize=normalize, ignore_index=ignore_index
+            ),
             metric_args={
                 "num_classes": NUM_CLASSES,
                 "normalize": normalize,
@@ -169,7 +171,9 @@ class TestMulticlassConfusionMatrix(MetricTester):
             preds=preds,
             target=target,
             metric_functional=multiclass_confusion_matrix,
-            reference_metric=partial(_sklearn_confusion_matrix_multiclass, normalize=normalize, ignore_index=ignore_index),
+            reference_metric=partial(
+                _sklearn_confusion_matrix_multiclass, normalize=normalize, ignore_index=ignore_index
+            ),
             metric_args={
                 "num_classes": NUM_CLASSES,
                 "normalize": normalize,
@@ -245,7 +249,9 @@ class TestMultilabelConfusionMatrix(MetricTester):
             preds=preds,
             target=target,
             metric_class=MultilabelConfusionMatrix,
-            reference_metric=partial(_sklearn_confusion_matrix_multilabel, normalize=normalize, ignore_index=ignore_index),
+            reference_metric=partial(
+                _sklearn_confusion_matrix_multilabel, normalize=normalize, ignore_index=ignore_index
+            ),
             metric_args={
                 "num_labels": NUM_CLASSES,
                 "normalize": normalize,
@@ -263,7 +269,9 @@ class TestMultilabelConfusionMatrix(MetricTester):
             preds=preds,
             target=target,
             metric_functional=multilabel_confusion_matrix,
-            reference_metric=partial(_sklearn_confusion_matrix_multilabel, normalize=normalize, ignore_index=ignore_index),
+            reference_metric=partial(
+                _sklearn_confusion_matrix_multilabel, normalize=normalize, ignore_index=ignore_index
+            ),
             metric_args={
                 "num_labels": NUM_CLASSES,
                 "normalize": normalize,
