+++ tests/unittests/image/test_ergas.py
@@ -13,14 +13,15 @@
 # limitations under the License.
 from collections import namedtuple
 from functools import partial
+from typing import Union
 
 import pytest
 import torch
+from torch import Tensor
 
 from torchmetrics.functional.image.ergas import error_relative_global_dimensionless_synthesis
 from torchmetrics.image.ergas import ErrorRelativeGlobalDimensionlessSynthesis
 from unittests.helpers import seed_all
-from unittests.helpers.reference_metrics import _baseline_ergas
 from unittests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
 
 seed_all(42)
@@ -36,6 +37,37 @@ for size, channel, coef, ratio, dtype in [
 ]:
     preds = torch.rand(NUM_BATCHES, BATCH_SIZE, channel, size, size, dtype=dtype)
     _inputs.append(Input(preds=preds, target=preds * coef, ratio=ratio))
+
+
+def _baseline_ergas(
+    preds: Tensor,
+    target: Tensor,
+    ratio: Union[int, float] = 4,
+    reduction: str = "elementwise_mean",
+) -> Tensor:
+    """Reference implementation of Erreur Relative Globale Adimensionnelle de Synthèse."""
+    reduction_options = ("elementwise_mean", "sum", "none")
+    if reduction not in reduction_options:
+        raise ValueError(f"reduction has to be one of {reduction_options}, got: {reduction}.")
+    # reshape to (batch_size, channel, height*width)
+    b, c, h, w = preds.shape
+    sk_preds = preds.reshape(b, c, h * w)
+    sk_target = target.reshape(b, c, h * w)
+    # compute rmse per band
+    diff = sk_preds - sk_target
+    sum_squared_error = torch.sum(diff * diff, dim=2)
+    rmse_per_band = torch.sqrt(sum_squared_error / (h * w))
+    mean_target = torch.mean(sk_target, dim=2)
+    # compute ergas score
+    ergas_score = 100 * ratio * torch.sqrt(torch.sum((rmse_per_band / mean_target) ** 2, dim=1) / c)
+    # reduction
+    if reduction == "sum":
+        to_return = torch.sum(ergas_score)
+    elif reduction == "elementwise_mean":
+        to_return = torch.mean(ergas_score)
+    else:
+        to_return = ergas_score
+    return to_return
 
 
 @pytest.mark.parametrize("reduction", ["sum", "elementwise_mean"])
